fix: delete router issue
