+++ app/api/user/route.js
@@ -60,55 +60,3 @@ export async function GET(requests) {
     headers: { "Content-Type": "application/json" },
   });
 }
-
-/**
- * @swagger
- * /api/user/delete/{id}:
- *  delete:
- *   tags: [Users]
- *   summary: Deletes user from database
- *   description: Returns deleted user from database
- *   parameters:
- *    - in: path
- *      name: id
- *      required: true
- *      description: id of the user that you want to delete
- *   responses:
- *    200:
- *      description: deletes user
- *      content:
- *        application/json:
- *          schema:
- *            type: object
- *            properties:
- *              message:
- *                type: string
- *                description: Status message to reflecting deletion successfully
- *                example: Successfully deleted user
- */
-export async function DELETE(request) {
-  const { id } = request.params;
-  console.log("id2", id);
-
-  try {
-    let user = await prisma.user.findUnique({
-      where: { id: parseInt(id, 10) },
-    });
-
-    if (!user) throw new Error("No user found with this id.");
-
-    await prisma.user.delete({ where: { id: parseInt(id, 10) } });
-
-    return new Response(
-      JSON.stringify({ message: "Successfully deleted user" }),
-      {
-        headers: { "Content-Type": "application/json" },
-      },
-    );
-  } catch (e) {
-    return new Response(JSON.stringify({ message: e.message }), {
-      headers: { "Content-Type": "application/json" },
-      status: 500,
-    });
-  }
-}
